requirements.txt
torch
torchvision
hydra-core
omegaconf
numpy
Pillow
opencv-python
tqdm
requests